+++ app/assets/javascripts/govuk_publishing_components/lib/initial-focus.js
@@ -2,10 +2,11 @@ window.GOVUK = window.GOVUK || {}
 window.GOVUK.Modules = window.GOVUK.Modules || {};
 
 (function (Modules) {
-  function InitialFocus () { }
+  function InitialFocus ($module) {
+    this.$module = $module
+  }
 
-  InitialFocus.prototype.start = function ($module) {
-    this.$module = $module[0]
+  InitialFocus.prototype.init = function () {
     this.$module.focus()
   }